server cam now be started and stopped
