+++ src/main.py
@@ -1,44 +1,11 @@
 from initialize import readScript, findScript, scriptContainsExec
+from functions.sendInstruction import sendInstruction
+from functions.server import startServer, stopServer
 import argparse
-import socket
 import shutil
 import yaml
 import json
 import os
-
-HEADER_LENGTH = 10
-SOCKET_PATH = '/tmp/taskZen.sock'
-
-def sendInstruction(instruction: str, *, verbose: bool = True) -> None:
-    """
-    A function that sends an instruction over a Unix socket connection and recieves a response.
-
-    Parameters:
-        - instruction (str): The instruction to be sent.
-        - verbose (bool, optional): Flag to indicate whether to print the response. Defaults to True.
-
-    Returns:
-        None
-    """
-    def receiveMessage(sock: socket.socket) -> str:
-        header = sock.recv(HEADER_LENGTH).decode('utf-8')
-        if not header:
-            return None
-        messageLength = int(header.strip())
-        return sock.recv(messageLength).decode('utf-8')
-
-    with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as s:
-        socketPath = SOCKET_PATH
-        s.connect(socketPath)
-        message = f"{len(instruction):<{HEADER_LENGTH}}" + instruction
-        s.sendall(message.encode())
-        
-        while True:
-            response = receiveMessage(s)
-            if response == 'end':
-                break
-            if verbose and response:
-                print(response)
 
 def checkConfig() -> None:
     configDir = os.getenv('XDG_CONFIG_HOME', default=os.path.expanduser('~/.config')) + '/taskZen/'
@@ -161,37 +128,11 @@ def main() -> None:
     elif args.command in ['server']:
         if args.start:
             print('Starting taskZen server')
-            try:
-                instruction = {
-                    'instruction': 'ping'
-                }
-                sendInstruction(json.dumps(instruction))
-                print('Server already running')
-            except (ConnectionRefusedError, FileNotFoundError):
-                from subprocess import DEVNULL
-                import subprocess
-                currentPath = f'{os.path.dirname(os.path.realpath(__file__))}'
-                if 'FLATPAK_ID' in os.environ:
-                    pythonExecutable = "/usr/bin/python"
-                else:
-                    pythonExecutable = os.path.join(os.getenv('VIRTUAL_ENV', ''), 'bin/python')
-                subprocess.Popen(
-                    [pythonExecutable, f'{currentPath}/executionServer.py'],
-                    start_new_session=True,
-                    stdout=DEVNULL,
-                    stderr=DEVNULL,
-                    stdin=DEVNULL
-                )
+            print(startServer()[1])
 
         elif args.kill:
             print('Killing taskZen server')
-            try:
-                instruction = {
-                    'instruction': 'kill'
-                }
-                sendInstruction(json.dumps(instruction))
-            except (ConnectionRefusedError, FileNotFoundError):
-                print('Server not running')
+            print(stopServer()[1])
 
         else:
             parserServer.print_help()
